safari-600.7-branch/LayoutTests/fast/loader/non-deferred-substitute-load-expected.txt

ALERT: Triggering bug through JS-alert
PASS
